+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 2.1.3 (under development)
+# 2.1.3 (14-apr-2017)
 - Improved `go-hello-test` archetype to generate [Intellij Idea Go plugin project structure](https://plugins.jetbrains.com/plugin/5047-go-language-golang-org-support-plugin)
 - Added flag `enforceGoPathToEnd` to enforce changing of folder list order in new generated GOPATH
 - Added list parameter `ldFlags` for `buildMojo` to define linker flags.